# specs/001-resume-upload-rating/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Resume Fit Evaluator API
  version: 0.3.0
  description: |
    Single-user FastAPI service. Maintains a library of baseline resume
    profiles, evaluates job descriptions against a selected profile with
    three-tier scoring, and tracks history. Uploaded PDFs are persisted on
    local disk under `${STORAGE_DIR}/profiles/<id>/<utc-ts>.pdf` and the
    absolute path is returned on the response.
    All responses follow the envelope `{ data, error, meta }`.
servers:
  - url: http://localhost:8000
    description: Local dev
...
paths:
...
  /api/evaluate:
    post:
      tags: [evaluate]
      summary: Score a single JD against the baseline (three-tier)
      description: |
        Scores the JD 0–100 and returns one of three statuses:
        - `ready_to_submit` (85+): no tailoring needed
        - `needs_tailoring` (60–84): durable tailoring queued
        - `skip` (<60): poor fit, skip_reason provided
        Duplicate JDs (same content hash) return the cached result.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateIn"
      responses:
        "200":
          description: Evaluation result (new or cached)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluateEnvelope"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/evaluate/bulk:
    post:
      tags: [evaluate]
      summary: Score multiple JDs in one request
      description: |
        Evaluates each JD against the baseline. Deduplicates by content hash —
        a JD already in the DB returns `cached: true` without an LLM call.
        Durable tailoring is triggered for each new `needs_tailoring` result.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkEvaluateIn"
      responses:
        "200":
          description: Bulk evaluation summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkEvaluateEnvelope"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/evaluate/by-listings:
    post:
      tags: [evaluate]
      summary: Score stored JD Database listings and link successful analyses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateByListingsIn"
      responses:
        "200":
          description: Per-listing evaluation summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluateByListingsEnvelope"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/evaluate/pending-listings:
    post:
      tags: [evaluate]
      summary: Score pending JD Database listings and link successful analyses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluatePendingListingsIn"
      responses:
        "200":
          description: Pending-listing evaluation summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluateByListingsEnvelope"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/callback:
    post:
      tags: [evaluate]
      summary: Deliver a generated resume from an external pipeline
      description: |
        Accepts a generated resume (e.g., from n8n or a CI pipeline) and stores
        it. Sets can_submit=true on the parent JobAnalysis.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CallbackIn"
      responses:
        "200":
          description: Resume stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdEnvelope"
        "404":
          $ref: "#/components/responses/Error"
...
  /api/history/{id}:
    get:
      tags: [history]
      summary: Get a single job analysis with generated resumes
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Job analysis detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryDetailEnvelope"
        "404":
          $ref: "#/components/responses/Error"
...
  /api/generated-resumes/{id}/pdf:
    get:
      tags: [history]
      summary: Generate if needed and download a tailored resume PDF
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/Error"
...
components:
  schemas:
    Meta:
      type: object
      properties:
        request_id: { type: string }
        cached: { type: boolean }
      additionalProperties: true
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    EvaluateIn:
      type: object
      required: [jd_text]
      properties:
        jd_text: { type: string }
        profile_id:
          type: integer
          nullable: true
          description: If omitted, the default profile is used, falling back to the latest profile.
    EvaluateOut:
      type: object
      properties:
        job_analysis_id: { type: string, format: uuid }
        tailoring_job_id: { type: string, format: uuid, nullable: true }
        tailoring_status:
          type: string
          enum: [queued, running, retry_wait, cancel_requested, cancelled, succeeded, failed]
          nullable: true
        score: { type: integer, minimum: 0, maximum: 100 }
        explanation: { type: string }
        strengths:
          type: array
          items: { type: string }
        gaps:
          type: array
          items: { type: string }
        threshold_met: { type: boolean }
        status:
          type: string
          enum: [ready_to_submit, needs_tailoring, skip]
        message: { type: string }
        action:
          type: string
          enum: [none, tailoring, skip]
        prompt_version: { type: string, nullable: true }
        llm_backend: { type: string, nullable: true }
        llm_model: { type: string, nullable: true }
    BulkEvaluateIn:
      type: object
      required: [jd_texts]
      properties:
        jd_texts:
          type: array
          items: { type: string }
          minItems: 1
          maxItems: 100
    BulkEvaluateResult:
      type: object
      properties:
        job_analysis_id: { type: string, format: uuid }
        tailoring_job_id: { type: string, format: uuid, nullable: true }
        tailoring_status:
          type: string
          enum: [queued, running, retry_wait, cancel_requested, cancelled, succeeded, failed]
          nullable: true
        jd_snippet: { type: string }
        score: { type: integer, minimum: 0, maximum: 100 }
        status:
          type: string
          enum: [ready_to_submit, needs_tailoring, skip]
        cached: { type: boolean }
    BulkEvaluateOut:
      type: object
      properties:
        total: { type: integer }
        new: { type: integer }
        cached: { type: integer }
        skipped: { type: integer }
        blocked: { type: integer }
        estimated_input_tokens: { type: integer, nullable: true }
        estimated_output_tokens: { type: integer, nullable: true }
        estimated_total_tokens: { type: integer, nullable: true }
        estimated_cost_usd: { type: number, nullable: true }
        results:
          type: array
          items: { $ref: "#/components/schemas/BulkEvaluateResult" }
    EvaluateByListingsIn:
      type: object
      required: [job_listing_ids]
      properties:
        profile_id: { type: integer, nullable: true }
        job_listing_ids:
          type: array
          minItems: 1
          maxItems: 20
          items: { type: string, format: uuid }
    EvaluatePendingListingsIn:
      type: object
      properties:
        profile_id: { type: integer, nullable: true }
        source: { type: string, nullable: true }
        limit: { type: integer, minimum: 1, maximum: 100, default: 100 }
    EvaluateByListingsResult:
      type: object
      properties:
        listing_id: { type: string, format: uuid }
        job_analysis_id: { type: string, format: uuid, nullable: true }
        tailoring_job_id: { type: string, format: uuid, nullable: true }
        tailoring_status:
          type: string
          enum: [queued, running, retry_wait, cancel_requested, cancelled, succeeded, failed]
          nullable: true
        score: { type: integer, minimum: 0, maximum: 100, nullable: true }
        status:
          type: string
          enum: [ready_to_submit, needs_tailoring, skip]
          nullable: true
        cached: { type: boolean }
        error: { type: string, nullable: true }
    EvaluateByListingsOut:
      type: object
      properties:
        total: { type: integer }
        succeeded: { type: integer }
        failed: { type: integer }
        skipped: { type: integer }
        blocked: { type: integer }
        tailoring_blocked: { type: integer }
        estimated_input_tokens: { type: integer, nullable: true }
        estimated_output_tokens: { type: integer, nullable: true }
        estimated_total_tokens: { type: integer, nullable: true }
        estimated_cost_usd: { type: number, nullable: true }
        warnings:
          type: array
          items:
            type: object
            additionalProperties: true
        results:
          type: array
          items: { $ref: "#/components/schemas/EvaluateByListingsResult" }
    CallbackIn:
      type: object
      required: [job_analysis_id, resume_text]
      properties:
        job_analysis_id: { type: string, format: uuid }
        resume_text: { type: string }
        pdf_url: { type: string, nullable: true }
        prompt_version: { type: string, nullable: true }
    HistoryItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        profile_id: { type: integer, nullable: true }
        tailoring_job_id: { type: string, format: uuid, nullable: true }
        tailoring_status:
          type: string
          enum: [queued, running, retry_wait, cancel_requested, cancelled, succeeded, failed]
          nullable: true
        jd_snippet: { type: string, nullable: true }
        score: { type: integer, nullable: true }
        threshold_met: { type: boolean }
        status: { type: string, nullable: true }
        can_submit: { type: boolean }
        prompt_version: { type: string, nullable: true }
        llm_backend: { type: string, nullable: true }
        llm_model: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
    HistoryDetail:
      allOf:
        - $ref: "#/components/schemas/HistoryItem"
        - type: object
          properties:
            jd_full_text: { type: string }
            skip_reason: { type: string, nullable: true }
            baseline_profile_text: { type: string, nullable: true }
            generated_resumes:
              type: array
              items: { $ref: "#/components/schemas/GeneratedResume" }
...
    EvaluateEnvelope:
      type: object
      properties:
        data: { $ref: "#/components/schemas/EvaluateOut" }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
    BulkEvaluateEnvelope:
      type: object
      properties:
        data: { $ref: "#/components/schemas/BulkEvaluateOut" }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
    EvaluateByListingsEnvelope:
      type: object
      properties:
        data: { $ref: "#/components/schemas/EvaluateByListingsOut" }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
...
    HistoryDetailEnvelope:
      type: object
      properties:
        data: { $ref: "#/components/schemas/HistoryDetail" }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
...
    IdEnvelope:
      type: object
      properties:
        data:
          type: object
          properties:
            id: { type: string, format: uuid }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
    ErrorEnvelope:
      type: object
      properties:
        data: { type: "null" }
        error: { $ref: "#/components/schemas/Error" }
        meta: { $ref: "#/components/schemas/Meta" }
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"